+++ clarity/src/vm/costs/cost_functions.rs
@@ -158,6 +158,7 @@ define_named_enum!(ClarityCostFunction {
     BitwiseLShift("cost_bitwise_left_shift"),
     BitwiseRShift("cost_bitwise_right_shift"),
     ContractHash("cost_contract_hash"),
+    ToAscii("cost_to_ascii"),
     Unimplemented("cost_unimplemented"),
 });
 
@@ -328,6 +329,7 @@ pub trait CostValues {
     fn cost_bitwise_left_shift(n: u64) -> InterpreterResult<ExecutionCost>;
     fn cost_bitwise_right_shift(n: u64) -> InterpreterResult<ExecutionCost>;
     fn cost_contract_hash(n: u64) -> InterpreterResult<ExecutionCost>;
+    fn cost_to_ascii(n: u64) -> InterpreterResult<ExecutionCost>;
 }
 
 impl ClarityCostFunction {
@@ -481,6 +483,7 @@ impl ClarityCostFunction {
             ClarityCostFunction::BitwiseLShift => C::cost_bitwise_left_shift(n),
             ClarityCostFunction::BitwiseRShift => C::cost_bitwise_right_shift(n),
             ClarityCostFunction::ContractHash => C::cost_contract_hash(n),
+            ClarityCostFunction::ToAscii => C::cost_to_ascii(n),
             ClarityCostFunction::Unimplemented => Err(RuntimeErrorType::NotImplemented.into()),
         }
     }