+++ clarity/src/vm/tests/variables.rs
@@ -55,7 +55,7 @@ fn test_block_height(
         let err = analysis.unwrap_err();
         assert_eq!(
             CheckErrors::UndefinedVariable("block-height".to_string()),
-            err.err
+            *err.err
         );
     } else {
         assert!(analysis.is_ok());
@@ -113,7 +113,7 @@ fn test_stacks_block_height(
         let err = analysis.unwrap_err();
         assert_eq!(
             CheckErrors::UndefinedVariable("stacks-block-height".to_string()),
-            err.err
+            *err.err
         );
     } else {
         assert!(analysis.is_ok());
@@ -173,7 +173,7 @@ fn test_tenure_height(
         let err = analysis.unwrap_err();
         assert_eq!(
             CheckErrors::UndefinedVariable("tenure-height".to_string()),
-            err.err
+            *err.err
         );
     } else {
         assert!(analysis.is_ok());
@@ -246,7 +246,7 @@ fn expect_contract_error(
     for (when, err_condition, expected_error) in expected_errors {
         if *when == WhenError::Analysis && err_condition(version, epoch) {
             let err = analysis.unwrap_err();
-            assert_eq!(*expected_error, err.err);
+            assert_eq!(*expected_error, *err.err);
 
             // Do not continue with the test if the analysis failed.
             return;
@@ -1150,7 +1150,7 @@ fn test_current_contract(
         let err = analysis.unwrap_err();
         assert_eq!(
             CheckErrors::UndefinedVariable("current-contract".to_string()),
-            err.err
+            *err.err
         );
     } else {
         assert!(analysis.is_ok());