+++ stackslib/src/clarity_vm/tests/contracts.rs
@@ -54,7 +54,7 @@ fn test_get_burn_block_info_eval() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::UnknownFunction(func_name) = check_error.err {
+                if let CheckErrors::UnknownFunction(func_name) = *check_error.err {
                     assert_eq!(func_name, "get-burn-block-info?");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);
@@ -79,7 +79,7 @@ fn test_get_burn_block_info_eval() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::UnknownFunction(func_name) = check_error.err {
+                if let CheckErrors::UnknownFunction(func_name) = *check_error.err {
                     assert_eq!(func_name, "get-burn-block-info?");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);
@@ -180,7 +180,7 @@ fn test_get_block_info_eval_v210() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::NoSuchBlockInfoProperty(name) = check_error.err {
+                if let CheckErrors::NoSuchBlockInfoProperty(name) = *check_error.err {
                     assert_eq!(name, "block-reward");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);
@@ -205,7 +205,7 @@ fn test_get_block_info_eval_v210() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::NoSuchBlockInfoProperty(name) = check_error.err {
+                if let CheckErrors::NoSuchBlockInfoProperty(name) = *check_error.err {
                     assert_eq!(name, "block-reward");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);
@@ -219,7 +219,7 @@ fn test_get_block_info_eval_v210() {
     sim.execute_next_block_as_conn(|conn| {
         let contract_identifier = QualifiedContractIdentifier::local("test-contract-3").unwrap();
         let contract =
-            "(define-private (test-func-1 (height uint)) (get-block-info? block-reward height)) 
+            "(define-private (test-func-1 (height uint)) (get-block-info? block-reward height))
              (define-private (test-func-2 (height uint)) (get-block-info? miner-spend-winner height))
              (define-private (test-func-3 (height uint)) (get-block-info? miner-spend-total height))";
         let epoch = conn.get_epoch();
@@ -249,7 +249,7 @@ fn test_get_block_info_eval_v210() {
             tx.eval_read_only(&contract_identifier, "(test-func-3 u0)")
                 .unwrap()
         );
-        // only works at the first block and later (not the 0th block) 
+        // only works at the first block and later (not the 0th block)
         assert_eq!(
             Value::some(Value::UInt(3000)).unwrap(),
             tx.eval_read_only(&contract_identifier, "(test-func-1 u1)")
@@ -367,7 +367,7 @@ fn trait_invocation_205_with_stored_principal() {
         let error = publish_contract(conn, &invoke_contract_id, invoke_contract, clarity_version)
             .unwrap_err();
         match error {
-            ClarityError::Analysis(ref e) => match e.err {
+            ClarityError::Analysis(ref e) => match *e.err {
                 CheckErrors::TypeError(..) => (),
                 _ => panic!("Unexpected error: {:?}", error),
             },
@@ -461,8 +461,8 @@ fn trait_invocation_cross_epoch() {
                 )
                 .unwrap_err();
 
-            if let ClarityError::Interpreter(Error::Unchecked(CheckErrors::TypeValueError(TypeSignature::TraitReferenceType(_), value))) = error {
-                // pass
+            if let ClarityError::Interpreter(Error::Unchecked(CheckErrors::TypeValueError(trait_ref_type, value))) = error {
+                assert!(matches!(*trait_ref_type, TypeSignature::TraitReferenceType(_)));
             } else {
                 panic!("Expected an Interpreter(UncheckedError(TypeValue(TraitReferenceType, Principal))) during Epoch-2.2");
             };
@@ -485,8 +485,8 @@ fn trait_invocation_cross_epoch() {
                 )
                 .unwrap_err();
 
-            if let ClarityError::Interpreter(Error::Unchecked(CheckErrors::TypeValueError(TypeSignature::TraitReferenceType(_), value))) = error {
-                // pass
+            if let ClarityError::Interpreter(Error::Unchecked(CheckErrors::TypeValueError(trait_ref_type, value))) = error {
+                assert!(matches!(*trait_ref_type, TypeSignature::TraitReferenceType(_)));
             } else {
                 panic!("Expected an Interpreter(UncheckedError(TypeValue(TraitReferenceType, Principal))) during Epoch-2.2");
             };
@@ -936,7 +936,7 @@ fn test_block_heights() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::UndefinedVariable(var_name) = check_error.err {
+                if let CheckErrors::UndefinedVariable(var_name) = *check_error.err {
                     assert_eq!(var_name, "stacks-block-height");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);
@@ -972,7 +972,7 @@ fn test_block_heights() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::UndefinedVariable(var_name) = check_error.err {
+                if let CheckErrors::UndefinedVariable(var_name) = *check_error.err {
                     assert_eq!(var_name, "stacks-block-height");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);
@@ -989,7 +989,7 @@ fn test_block_heights() {
                 ASTRules::PrecheckSize,
             );
             if let Err(ClarityError::Analysis(check_error)) = res {
-                if let CheckErrors::UndefinedVariable(var_name) = check_error.err {
+                if let CheckErrors::UndefinedVariable(var_name) = *check_error.err {
                     assert_eq!(var_name, "block-height");
                 } else {
                     panic!("Bad analysis error: {:?}", &check_error);